Upgrading RHVH hyperconverged host from Admin portal fails
Upgrading RHVH hyperconverged host from Admin portal
fails for first host

Upgrade is failing because ovit-ha-agent is trying to connect vdsm,
and if connection fails, it exits with a non-zero exit code.

after it fails systemd tries to restart the agent, and sincie the
agent wants vdsmd, it also starts vdsmd. This means that if you do:
   systemctl stop vdmsd
and ovirt-ha-agent is in the middle of restart loop,
vdsmd service will start again. This is an impossible situation
so hence just before upgrade stopping the agent and
broker service.

Bug-Url: https://bugzilla.redhat.com/1979539